+++ src/index.css
@@ -114,16 +114,16 @@
   --color-green-dark-active: #124e4b;
   --color-green-darker: #0e3d3a;
 
-  --color-orange-light: #fff5eb;
-  --color-orange-light-hover: #fff0e0;
-  --color-orange-light-active: #fee0bf;
-  --color-orange-normal: #fc9c32;
-  --color-orange-normal-hover: #e38c2d;
-  --color-orange-normal-active: #ca7d28;
-  --color-orange-dark: #bd7526;
-  --color-orange-dark-hover: #975e1e;
-  --color-orange-dark-active: #714616;
-  --color-orange-darker: #583712;
+  --color-yellow-light: #fef4e8;
+  --color-yellow-light-hover: #feefdd;
+  --color-yellow-light-active: #fddeb9;
+  --color-yellow-normal: #f7941d;
+  --color-yellow-normal-hover: #de851a;
+  --color-yellow-normal-active: #c67617;
+  --color-yellow-dark: #b96f16;
+  --color-yellow-dark-hover: #945911;
+  --color-yellow-dark-active: #6f430d;
+  --color-yellow-darker: #56340a;
 
   --color-grey-light: #ebebeb;
   --color-grey-light-hover: #e1e1e1;
